refactor(integration-tests): outsourced clients
